Catalogo de colonias busca por codigo postal y renderiza la tabla, navegacion y validacion

+++ tecolotl-web/src/main/java/tecolotl/web/control/administracion/EscuelaDetalleControlador.java
@@ -34,7 +34,7 @@ public class EscuelaDetalleControlador implements Serializable {
 
     private EscuelaDetalleDto escuelaDetalleDto;
     private ContactoDto contactoDtoModelo;
-    private int tipoContacto;
+    private short tipoContacto;
     private List<TipoContactoDto> tipoContactoDtoLista;
 
     @PostConstruct
@@ -79,11 +79,11 @@ public class EscuelaDetalleControlador implements Serializable {
         this.contactoDtoModelo = contactoDtoModelo;
     }
 
-    public int getTipoContacto() {
+    public short getTipoContacto() {
         return tipoContacto;
     }
 
-    public void setTipoContacto(int tipoContacto) {
+    public void setTipoContacto(short tipoContacto) {
         this.tipoContacto = tipoContacto;
     }
